Use heroku/authenticater instead
So we can share with other apps
