+++ Godeps/_workspace/src/github.com/heroku/authenticater/README.md
@@ -1,0 +1,9 @@
+# What
+
+Authenticator interface for Go HTTP apps.
+
+# Usage
+
+See [godoc](http://godoc.org/github.com/heroku/authenticater) or [examples](https://github.com/heroku/authenticater/tree/master/examples)
+
+
