+++ Godeps/_workspace/src/github.com/heroku/authenticater/any.go
@@ -1,0 +1,9 @@
+package authenticater
+
+import "net/http"
+
+type AnyOrNoAuth struct{}
+
+func (fa AnyOrNoAuth) Authenticate(r *http.Request) bool {
+	return true
+}
